+++ pkg/prober/prober_test.go
@@ -1,0 +1,162 @@
+package prober
+
+import (
+	"context"
+	"fmt"
+	"net"
+	"net/http"
+	"path"
+	"testing"
+	"time"
+
+	"github.com/go-kit/kit/log"
+	"github.com/improbable-eng/thanos/pkg/runutil"
+	"github.com/improbable-eng/thanos/pkg/testutil"
+	"github.com/oklog/run"
+	"github.com/pkg/errors"
+	"github.com/prometheus/common/route"
+)
+
+func queryHTTPGetEndpoint(ctx context.Context, t *testing.T, logger log.Logger, url string) (*http.Response, error) {
+	req, err := http.NewRequest("GET", fmt.Sprintf("http://%s", url), nil)
+	testutil.Ok(t, err)
+	return http.DefaultClient.Do(req.WithContext(ctx))
+}
+
+func TestProberHealthInitialState(t *testing.T) {
+	component := "test"
+	expectedErrorMessage := fmt.Sprintf(initialErrorText, component)
+	p := NewProber(component, log.NewNopLogger())
+
+	err := p.IsHealthy()
+	testutil.NotOk(t, err)
+	testutil.Equals(t, err.Error(), expectedErrorMessage)
+}
+
+func TestProberReadinessInitialState(t *testing.T) {
+	component := "test"
+	expectedErrorMessage := fmt.Sprintf(initialErrorText, component)
+	p := NewProber(component, log.NewNopLogger())
+
+	err := p.IsReady()
+	testutil.NotOk(t, err)
+	testutil.Equals(t, err.Error(), expectedErrorMessage)
+}
+
+func TestProberReadyStatusSetting(t *testing.T) {
+	component := "test"
+	testError := fmt.Errorf("test error")
+	p := NewProber(component, log.NewNopLogger())
+
+	p.SetReady()
+	err := p.IsReady()
+	testutil.Equals(t, err, nil)
+	p.SetNotReady(testError)
+	err = p.IsReady()
+	testutil.NotOk(t, err)
+}
+
+func TestProberHeatlthyStatusSetting(t *testing.T) {
+	component := "test"
+	testError := fmt.Errorf("test error")
+	p := NewProber(component, log.NewNopLogger())
+
+	p.SetHealthy()
+	err := p.IsHealthy()
+	testutil.Equals(t, err, nil)
+	p.SetNotHealthy(testError)
+	err = p.IsHealthy()
+	testutil.NotOk(t, err)
+}
+
+func TestProberMuxRegistering(t *testing.T) {
+	component := "test"
+	ctx := context.Background()
+	var g run.Group
+	mux := http.NewServeMux()
+
+	freePort, err := testutil.FreePort()
+	testutil.Ok(t, err)
+	serverAddress := fmt.Sprintf("localhost:%d", freePort)
+
+	l, err := net.Listen("tcp", serverAddress)
+	testutil.Ok(t, err)
+	g.Add(func() error {
+		return errors.Wrap(http.Serve(l, mux), "serve probes")
+	}, func(error) {})
+
+	p := NewProbeInMux(component, mux, log.NewNopLogger())
+
+	go func() { _ = g.Run() }()
+
+	testutil.Ok(t, runutil.Retry(time.Second, ctx.Done(), func() error {
+		resp, err := queryHTTPGetEndpoint(ctx, t, log.NewNopLogger(), path.Join(serverAddress, healthyEndpointPath))
+		testutil.Ok(t, err)
+		testutil.Equals(t, probeErrorHTTPStatus, resp.StatusCode)
+		return err
+	}))
+
+	testutil.Ok(t, runutil.Retry(time.Second, ctx.Done(), func() error {
+		resp, err := queryHTTPGetEndpoint(ctx, t, log.NewNopLogger(), path.Join(serverAddress, readyEndpointPath))
+		testutil.Ok(t, err)
+		testutil.Equals(t, probeErrorHTTPStatus, resp.StatusCode)
+		return err
+	}))
+
+	p.SetHealthy()
+
+	testutil.Ok(t, runutil.Retry(time.Second, ctx.Done(), func() error {
+		resp, err := queryHTTPGetEndpoint(ctx, t, log.NewNopLogger(), path.Join(serverAddress, healthyEndpointPath))
+		testutil.Ok(t, err)
+		testutil.Equals(t, 200, resp.StatusCode)
+		return err
+	}))
+
+}
+
+func TestProberRouterRegistering(t *testing.T) {
+	component := "test"
+	router := route.New()
+	ctx := context.Background()
+	var g run.Group
+	mux := http.NewServeMux()
+
+	freePort, err := testutil.FreePort()
+	testutil.Ok(t, err)
+	serverAddress := fmt.Sprintf("localhost:%d", freePort)
+
+	l, err := net.Listen("tcp", serverAddress)
+	testutil.Ok(t, err)
+	g.Add(func() error {
+		return errors.Wrap(http.Serve(l, mux), "serve probes")
+	}, func(error) {})
+
+	p := NewProbeInRouter(component, router, log.NewNopLogger())
+	mux.Handle("/", router)
+
+	go func() { _ = g.Run() }()
+
+	testutil.Ok(t, runutil.Retry(time.Second, ctx.Done(), func() error {
+		resp, err := queryHTTPGetEndpoint(ctx, t, log.NewNopLogger(), path.Join(serverAddress, healthyEndpointPath))
+		testutil.Ok(t, err)
+		testutil.Equals(t, probeErrorHTTPStatus, resp.StatusCode)
+		return err
+	}))
+
+	testutil.Ok(t, runutil.Retry(time.Second, ctx.Done(), func() error {
+		resp, err := queryHTTPGetEndpoint(ctx, t, log.NewNopLogger(), path.Join(serverAddress, readyEndpointPath))
+		testutil.Ok(t, err)
+		testutil.Equals(t, probeErrorHTTPStatus, resp.StatusCode)
+		return err
+	}))
+
+	p.SetHealthy()
+
+	testutil.Ok(t, runutil.Retry(time.Second, ctx.Done(), func() error {
+		resp, err := queryHTTPGetEndpoint(ctx, t, log.NewNopLogger(), path.Join(serverAddress, healthyEndpointPath))
+		testutil.Ok(t, err)
+		testutil.Equals(t, 200, resp.StatusCode)
+		return err
+	}))
+
+}